Refactor #2

+++ Drawer/Drawer/DeleteShapeCommand.h
@@ -5,13 +5,13 @@ class CDeleteShapeCommand :
 	public IShapeCommand
 {
 public:
-	CDeleteShapeCommand(CDrawerDoc* doc, ShapeType type, const Gdiplus::Rect& rect, int delIndex);
+	CDeleteShapeCommand(CDrawerDoc* doc, ShapeType type, const Gdiplus::Rect& rect, size_t delIndex);
 	void Undo() override;
 	void Redo() override;
 
 private:
 	ShapeType m_type;
 	Gdiplus::Rect m_rect;
-	int m_delIndex;
+	size_t m_delIndex;
 };
